Tests/LibWeb/Text/expected/URL/search-params-with-no-query.txt

https://www.birdoftheyear.org.nz/?
https://www.birdoftheyear.org.nz/
